Using last Nodes Inset and display
Circular linked list is a linked list where all nodes are connected to form a circle. There is no NULL at the end. A circular linked list can be a singly circular linked list or doubly circular linked list.



Advantages of Circular Linked Lists:
1) Any node can be a starting point. We can traverse the whole list by starting from any point. We just need to stop when the first visited node is visited again.

2) Useful for implementation of queue. Unlike this implementation, we don’t need to maintain two pointers for front and rear if we use circular linked list. We can maintain a pointer to the last inserted node and front can always be obtained as next of last.

3) Circular lists are useful in applications to repeatedly go around the list. For example, when multiple applications are running on a PC, it is common for the operating system to put the running applications on a list and then to cycle through them, giving each of them a slice of time to execute, and then making them wait while the CPU is given to another application. It is convenient for the operating system to use a circular list so that when it reaches the end of the list it can cycle around to the front of the list.

4) Circular Doubly Linked Lists are used for implementation of advanced data structures like

Implementation 
To implement a circular singly linked list, we take an external pointer that points to the last node of the list. If we have a pointer last pointing to the last node, then last -> next will point to the first node. 





The pointer last points to node Z and last -> next points to node P.

Why have we taken a pointer that points to the last node instead of first node ? 
For insertion of node in the beginning we need traverse the whole list. Also, for insertion at the end, the whole list has to be traversed. If instead of start pointer we take a pointer to the last node then in both the cases there won’t be any need to traverse the whole list. So insertion in the beginning or at the end takes constant time irrespective of the length of the list.

Insertion 
A node can be added in three ways: 

Insertion in an empty list
Insertion at the beginning of the list
Insertion at the end of the list
Insertion in between the nodes
Insertion in an empty List 
Initially, when the list is empty, last pointer will be NULL. 
 



After inserting a node T, 
 





After insertion, T is the last node so pointer last points to node T. And Node T is first and last node, so T is pointing to itself. 
Function to insert node in an empty List, 
  
Insertion at the beginning of the list 
To Insert a node at the beginning of the list, follow these step: 
1. Create a node, say T. 
2. Make T -> next = last -> next. 
3. last -> next = T. 

  
Insertion at the end of the list 
To Insert a node at the end of the list, follow these step: 
1. Create a node, say T. 
2. Make T -> next = last -> next; 
3. last -> next = T. 
4. last = T. 

Insertion in between the nodes 
To Insert a node in between the two nodes, follow these step: 
1. Create a node, say T. 
2. Search the node after which T need to be insert, say that node be P. 
3. Make T -> next = P -> next; 
4. P -> next = T.
Suppose 12 need to be insert after node having value 10,