tortoise and hare algorithm
You can see that we have used just one while loop in the getLastNode(int n) method. This method accepts an integer parameter and can be used to find the nth element from the end in the linked list like to get the 2nd element from the tail, pass 2 and to get the 3rd element from the list, pass 3. Similarly to get the Kth node from the tail, just pass K.

The SinglyLinkedList class represents a singly linked list in Java. It is the same class that we have used earlier linked list articles like writing your own linked list in Java. It is a collection of Node class which represents a node in the linked list. It contains a data part and a reference to the next node. The SinglyLinkedList class also contains a reference to the head, the first node of the linked list.

As a programmer, you should know key data structures and algorithms like what is a linked list data structure, it's pros and cons, and when to use a linked list like it is good for frequently adding and removing element but not so good for search as it takes O(n) time to find an element inside the linked list.

You can read more about the linked list in a good data structure and algorithm course