Create find length of linkedlist using recurstion
This is a much more concise then Iterative solution, but sometimes it's not easy to come up with the answer if you are not familiar with Recursion, but if you do, then it's straightforward.

public int length(Node current){
 if(current == null){ //base case
   return 0;
 }
 return 1+length(current.next());
}

You can see that we have used the fact that the last node will point to null to terminate the recursion. This is called the base case. It's essential to identify a base case while coding a recursive solution, without a base case, your program will never terminate and result in StackOverFlowError.